Preparing to release v2 packages w/ the v1 module.
This might be a terrible idea. My thinking is that I'm not ready to
commit to a v2 release yet, but I want to start using the v2 packages.
So, release them as part of v1, so the go.mod uses:

    require github.com/mdwhatcott/funcy v1.9.0

but import statments can use:

    import "github.com/mdwhatcott/funcy/v2"

When I'm comfy with the v2 packages it's a simple matter of putting a
`go.mod` in the /v2 folder and stamping a `v2.*.*` tag.

But doing this does mean that the /v2 packages are now part of the v1
contract (as of whatever version I release next).